factor out input-modes
keymaps now have a property 'display_name', from which information about
the currently active keymaps can be gotten for display in the minibuffer
area.  the ui is slightly more informative than before, because now it
will also tell you such information as whether the current context is in a
form.

content_buffer_modality has also been changed to give a better way for
page-modes to install their keymaps in the proper order in the keymap
stack.  references to keymaps are almost completely factored out of the
conkeror core now, which brings us one step closer to having alternative
binding sets which may be entirely different from the default binding set.