buffer (re)ordering
This patch introduces a design for the spatial organization (left to
right) of buffers, a canonical order independent of (but not in conflict
with) the temporal ordering (based on last-access time) of buffers that
was there before.  Spatial ordering of buffers lends itself especially
well to a tabs-based UI or other on-screen visualization of the
buffer-list.  Having both temporal and spatial ordering of buffers within
the same system raises a number of problems, which are most plainly
illustrated by the question, what should the 'bury-buffer' command do?
'Bury' ought to do something in the spatial ordering as well as the
temporal one, but what?  To reconcile the differences, we change the
default ordering of buffers from the old way of always putting new buffers
on the end of the buffer list to this new system:

New buffers opened by url or webjump are opened at the front of the buffer
list.  New buffers opened by following a link or other dom element are
inserted into the buffer list immediately to the right of the origin
buffer.  Bury-buffer moves a buffer all the way to the end of the buffer
list, and all the way to the end of the access-time list.  The end of the
(spatial) list is thus considered to be "buried" position, and this is why
we do not open new buffers in that position anymore.

Buffers can be manually re-ordered.

buffer-move-forward, buffer-move-backward: new commands, bound to M-N and M-P

move_buffer_hook: new hook, called when a buffer is moved (or buried).

buffer_container.buffer_list is now a list of the canonical ordering of
buffers.  last-accessed order is now tracked in buffer_container.buffer_history.

tab-bar.js, new-tabs.js, buffer_count_widget, session_load,
session-auto-save updated.

buffer: new keyword $position.  when its value is "end", the buffer is
added to the end of the buffer list rather than the default position.
This mechanism is used by session_load to restore the session buffers in
the correct order, and also admits for possible future expansion to allow
more user configuration of default buffer positions.

This patch is partly based on the work of Ben Spencer and Peter Lunicks.