browser_object_follow: do not give $opener for following non-element/window in new buffer
The rationale behind this change is that the current buffer is not, literally
speaking, the "opener" when the user has typed an url or webjump to open via
a command like find-url-new-buffer.

This change also prepares the way for the forthcoming implementation of
buffer-ordering, wherein the $opener of a buffer is used to determine where
to insert the new buffer in the window's buffer_list.