configurable default buffer ordering
new_buffer_position, new_buffer_with_opener_position: new user variables.

These two user variables control the default position of new buffers,
without and with an opener (in the current window), respectively.  Allowed
values are 'before', 'after', 'end' or an integer.  'before' and 'after'
are relative to the opener or current buffer.  'end' means the end of the
list.  An integer means an explicit position, with 0 giving the front of
the list.